Change Integer to BigInteger

+++ src/main/java/org/telosys/tools/dsl/parser/model2/DomainEnumerationForInteger.java
@@ -1,5 +1,6 @@
 package org.telosys.tools.dsl.parser.model2;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -7,29 +8,29 @@ import java.util.List;
  * @date 2014-05-13
  * @version 1.0
  */
-public class DomainEnumerationForInteger extends DomainEnumeration<Integer> {
+public class DomainEnumerationForInteger extends DomainEnumeration<BigInteger> {
 
     public DomainEnumerationForInteger(String name) {
     	super(name);
     }
 
-    public void addItem( String name, Integer value) {
+    public void addItem( String name, BigInteger value) {
         super.addItem(name, value);
     }
 
-    public void addItem(DomainEnumerationItem<Integer> item) {
+    public void addItem(DomainEnumerationItem<BigInteger> item) {
         super.addItem(item);
     }
 
-    public DomainEnumerationItem<Integer> getItem(String name) {
+    public DomainEnumerationItem<BigInteger> getItem(String name) {
     	return super.getItem(name);
     }
 
-    public Integer getItemValue(String name) {
+    public BigInteger getItemValue(String name) {
     	return super.getItemValue(name);
     }
 
-    public List<DomainEnumerationItem<Integer>> getItems() {
+    public List<DomainEnumerationItem<BigInteger>> getItems() {
         return super.getItems();
     }
